Added checkers in backend side and in front login, need to implement rest of frontend

+++ BackendSpring/backend/src/main/java/org/unibl/etf/ip/backend/controller/FitnessProgramController.java
@@ -4,14 +4,12 @@ package org.unibl.etf.ip.backend.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
+import org.unibl.etf.ip.backend.errorservice.ForbiddenEntity;
 import org.unibl.etf.ip.backend.exceptions.MethodNotAllowedException;
 import org.unibl.etf.ip.backend.exceptions.NotEnoughMoneyException;
 import org.unibl.etf.ip.backend.exceptions.NotFoundException;
-import org.unibl.etf.ip.backend.model.KorisnikEntity;
+import org.unibl.etf.ip.backend.loginservice.UserLoginHelp;
 import org.unibl.etf.ip.backend.model.KorisnikPretplacenProgramEntity;
 import org.unibl.etf.ip.backend.model.ProgramEntity;
 import org.unibl.etf.ip.backend.service.FitnessProgramService;
@@ -36,34 +34,34 @@ public class FitnessProgramController {
         return new ResponseEntity<>(service.getProgramById(id), HttpStatus.OK);
     }
 
-    //@PreAuthorize("hasRole('FITNESS_USER')")
     @GetMapping("/my-programs/{id}")
     public ResponseEntity<List<ProgramEntity>> getMyPrograms(@PathVariable("id") Integer id) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
-        // Extract the username and any other details from the Authentication object
-        String username = authentication.getName();
-        System.out.println("username" + username);
-        // Assuming you have a custom UserDetails implementation with an "id" property
-        KorisnikEntity userDetails = (KorisnikEntity) authentication.getPrincipal();
-        Integer userId = userDetails.getId();
-        System.out.println("ID" + userId);
+        if(!UserLoginHelp.checkUserValidity(id)) {
+            return ForbiddenEntity.returnForbidden();
+        }
+
         return new ResponseEntity<>(service.getMyPrograms(id), HttpStatus.OK);
     }
 
     @PostMapping
     public ResponseEntity<ProgramEntity> createProgram(@RequestBody ProgramEntity fitnessProgram) {
-        try {
-            return new ResponseEntity<>(service.createProgram(fitnessProgram), HttpStatus.OK);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        Integer userId = fitnessProgram.getKreatorId();
+        if(!UserLoginHelp.checkUserValidity(userId)) {
+            return ForbiddenEntity.returnForbidden();
         }
+
+        return new ResponseEntity<>(service.createProgram(fitnessProgram), HttpStatus.OK);
     }
 
     @DeleteMapping("/{programId}/{userId}")
     public ResponseEntity<String> deleteProgram(@PathVariable("programId") Integer programId,
                                                 @PathVariable("userId") Integer userId) throws NotFoundException, MethodNotAllowedException {
+        if(!UserLoginHelp.checkUserValidity(userId)) {
+            return ForbiddenEntity.returnForbidden();
+        }
+
+
         service.deleteProgram(programId, userId);
         return new ResponseEntity<>("Success", HttpStatus.OK);
     }
@@ -71,21 +69,38 @@ public class FitnessProgramController {
     @PostMapping("/subscribe")
     public ResponseEntity<KorisnikPretplacenProgramEntity> subscribeToAProgram(@RequestBody KorisnikPretplacenProgramEntity subscribeEntity)
             throws NotFoundException, NotEnoughMoneyException {
+
+        Integer userId = subscribeEntity.getKorisnikId();
+        if(!UserLoginHelp.checkUserValidity(userId)) {
+            return ForbiddenEntity.returnForbidden();
+        }
+
         return new ResponseEntity<>(service.subscribeToAProgram(subscribeEntity), HttpStatus.OK);
     }
 
     @GetMapping("/upcoming-user-programs/{userId}")
     public ResponseEntity<List<ProgramEntity>> getUserParticipations(@PathVariable("userId")Integer userId) {
+        if(!UserLoginHelp.checkUserValidity(userId)) {
+            return ForbiddenEntity.returnForbidden();
+        }
+
         return new ResponseEntity<>(service.getUserParticipations(userId), HttpStatus.OK);
     }
 
     @GetMapping("/past-user-programs/{userId}")
     public ResponseEntity<List<ProgramEntity>> getPastUserParticipations(@PathVariable("userId")Integer userId) {
+        if(!UserLoginHelp.checkUserValidity(userId)) {
+            System.out.println("Not same");
+            return ForbiddenEntity.returnForbidden();
+        }
         return new ResponseEntity<>(service.getPastUserParticipations(userId), HttpStatus.OK);
     }
 
     @GetMapping("/unparticipated-user-programs/{userId}")
     public ResponseEntity<List<ProgramEntity>> getUserUnparticipated(@PathVariable("userId")Integer userId) {
+        if(!UserLoginHelp.checkUserValidity(userId)) {
+            return ForbiddenEntity.returnForbidden();
+        }
         return new ResponseEntity<>(service.getUserUnparticipations(userId), HttpStatus.OK);
     }
